pyproject.toml
[project]
name = "douyin-collector"
version = "1.0.0"
description = "抖音收藏视频采集和上传工具"
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.48.0",
    "httpx>=0.25.0",
    "pyyaml>=6.0",
    "loguru>=0.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = []
